High-contrast hover style for date and week picker should be outline instead of shading the cells
I got mixed up when looking at the high contrast results for
crrev.com/c/2083866 and thought that shading the hovered cells was
correct.  Instead, the old behavior for date/week as well as month,
time, etc was to just draw an border for hover in high contrast mode.

This change restores the old behavior of just drawing a border on
hover for date and week in high contrast mode by mirroring the rules
for these controls in the forced-colors block.

Month wasn't broken by this because my change didn't affect the structure
of its highlight rule in the same way that it did for date and week.

Change-Id: I20fc850a782ade1170ea05653e454d827a59589c
Bug: 1062017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2103727
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752195}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9088f449b5a2e65bd4bb516213d642727e237838