FontAccess: Experimental Font Enumeration API
This implements an experimental async iterator-based API to
enumerate installed unique fonts on a system, available on Frames
and Workers.

This is a new browser capability that aims to provide web
developers low-level access to fonts. The first step is to create
the ability to enumerate fonts present on the system.

This access could be used to obtain the ability to control how
text is rendered, with the addition of Font Table Access, in an
upcoming changelist.

The backend this API gets its data from is implemented in
FontCache.

This is the first iteration of this API. It only works on Mac
at the moment.

Bug: 1043306
Change-Id: I5185c806e3db33ab6c5bdd0604e2d1ece8e78e43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106747
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Olivier Yiptong <oyiptong@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752231}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bbe7d4ee798837f9d252bb9138d3a77ecb9d0030