Expose the <html> element to the accessibility tree.
Last year we made a change to include some accessibility
objects in our internal tree even though they're "ignored".
However, this leads to a weird state where some ignored
nodes are included, but others aren't included.

To simplify things, we should include all accessibility
objects in the tree unless we prune an entire subtree.
We should never have a case where one node isn't included
but its children are, because then changing it triggers
some annoying reparenting.

This change is one step in that direction: it adds a
feature flag that exposes the <html> element. The
HTML element is always ignored, but to simplify the
logic we want it always included in the tree.

Follow-up changes will enable this flag for more and
more tests until we can just turn it on completely.
Then we can follow up with more changes until finally
we'll no longer have any nodes that are ignored and
excluded from the tree.

Bug: 1063155

Change-Id: I3983fa0f84244448dea9274fb2317bdaf2675869
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107142
Reviewed-by: Adam Ettenberger <Adam.Ettenberger@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752272}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8cc4152ab727b2cb8046eee0c396f348c9e21ad6