Fix Canvas2D perf regression from  "Add a PaintRecorder to CanvasResourceProvider"
This patch addresses a performance regressions introduced by
https://chromium.googlesource.com/chromium/src/+/d9b56fbb6c14e83389dfff002d43ff06edacdde8

The root cause here was performance hit caused by additional page faults
that were occurring because the CanvasResourceProvider was unnecessarily
holding on to the last recording.

The Canvas2DLayerBridge::FlushRecording would have previously cleared
the recording right away inside

if (!clear_frame_ || !resource_host_ || !resource_host_->IsPrinting()) {
    last_recording_ = nullptr;
    clear_frame_ = false;
}

because we moved the recording to the resource provider this was not
happening anymore.

Fix is to not hold on to the recording in the resource provider.

Change-Id: I6e4c76bc18bc685dd24843c2ee0d73d9f6a72f8a
Bug: 1056704
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112967
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf9e0a9161fcb675be02a17638d73f15d606785b