Fix a typo in InputMethodSettings.
This is a follow-up cl of https://crrev.com/c/2099343.
Make the naming and description of zhuyinKeyboardLayout and
koreanKeyboardLayout in a same fashion.

This class is not used in production yet, so it is still safe to change
field names.

Test: tested with Linux emulator
Bug: 1060906
Change-Id: I3305fd54a76e6f48ec6d3c3f4fb50b4cac7ad167
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111811
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Jing Wang <jiwan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752348}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 78436982f2a3dfc16da4e860417498185199dbdc