@page size descriptor with orientation only does not have a size.
The print preview UI used HasCustomPageSizeStyle() to decide if the CSS
dictates the given page size, or if a UI to change the page size should
be available when printing to PDF.

We returned true when the size descriptor only contained an orientation,
which does not give the user an opportunity to select if "size:landscape"
should print to Letter or A4.

Instead, replace HasCustomPageSizeStyle() with a GetPageSizeType() to
pass the page size type out of Blink instead of a boolean to make it
possible for the printing UI to disable the UI for selecting page size
and orientation separately.

The testRunner API for hasCustomPageSizeStyle() is removed since it was
only used for a single test which we now have coverage for in unit tests.

This CL should not have any behavioral changes.

Bug: 1016235
Change-Id: Id12c4310f87fc99b17660c355913150f12ff3e93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2107598
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752392}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e22d298b224ba358446e6cbba628a67ae2cfc71