IDL compier: Make it possible to compare String and IDL enumeration
While we migrate the code generator of IDL bindings, we need the feature
to compare IDL enumeration with String.  This CL implements it.
This CL affects on the new code generator, which is enabled behind a
flag, and not on the product.


Bug: 839389
Change-Id: I99e8f458f42decae42552b2c3fb998067234b1df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2115861
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752739}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2484902860a0401eef31808eba9b8a38c9a2238d