IDL Compiler: Enable to assign IDL Enum with String
While we migrate the code generator of IDL bindings, we need assignments
of enumeration with String.  This CL adds it as a migration adapter.
The CL affects on the new code generator, which is enabled behind a
flag, and not on the product.


Bug: 839389
Change-Id: I81f82678da5e82a4a25475d5c1707e00eefe82b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2116157
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752753}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4af655f5b5cd9042fde6b0dae922790351309382