Make the Portals RuntimeEnabledFeature triggered by an origin trial feature name.
This means that it can vary by context, so the last remaining sites that
did not include a FeatureContext (in practice, a Document) are updated
to do so.

So that force-enabling portals at the command line continues to make
it available everywhere, a mode which propagates the feature to the
Blink REF only if it's forced on at the command line (or with
chrome://flags) is added to content/child/.

On the flipside, having the REF turned on by the origin trial system
without the browser side enabled is nonsensical and the browser will
(rightfully) kill the renderer for attempting it. This adds a hook
next to the existing OS-check hook for specific logic. This could be
made more abstract, but at present we're the only known use, and the
relevant check is of a feature exposed in blink/public/common/, so
this seems fine for now.

Bug: 1040627
Change-Id: I75b31b5131d7513ec2bcf609f2ac34c24df8003f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2098986
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752873}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c39131b385fe661c82a1045d2659fc1648b9c10a