Report FirstInputDelay even if it's less than 1ms
Currently, the Blink methods return an unsigned integer, which is 0 when
the value has not yet been computed. This is a problem in cases when the
value is computed but very small because the integer is in ms. This CL
fixes that. To show that it works, this CL relands the  browser test for
LCP stopping upon input which was originally in
https://chromium-review.googlesource.com/c/chromium/src/+/2101541

Bug: 1050727, 936867
Change-Id: I756dfc9a902b033af4d2185362581de75c2c54d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112983
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74dcc97f37e8eaff2482fdf36a2be00164ae8253