Don't update paint invalidation container in CAP
Paint invalidation container information is no available in CAP, as it
requires querying the compositing state to see whether a GraphicsLayer
is created in a PaintLayer's ancestor chain.

UpdatePaintInvalidationContainer should therefore be a no-op in CAP.
This is an incremental change in pursuit of adding a DCHECK(!CAP) in
LayoutObject::EnclosingCompositedContainer and removing that
DisableCompositingQueryAssert from CAP codepaths.

Bug: 1007989
Change-Id: I0ecd12fdf907f1746d54a67437143cb5ce7eb5fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118212
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#753061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 689603b082bc7ead39591a577fe9b4fa6706e7fd