SubtreeVisibility: turn off script focus activation + tests.
This patch adds the ability for us to keep auto elements unlocked
when the subtree has focus.

The activation flag is replaced by an array of flags, and a function
uses those flags to determine whether we should lock or unlock.

This also adds a bunch of tests for focus().

Note that the place the focus/display lock updates happen needs
to before the scroll, and NotifyFocusedElementChanged is too
late in the pipeline for this.

Change-Id: I251dcca44275322d4ba71b7de6f8fb83eb1fe81c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2112956
Commit-Queue: vmpstr <vmpstr@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#753243}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68f4208d1b7a177e3c9ddee5fb87ccb663f492a6