Make IsFeatureEnabled(DocumentPolicyFeature...) const qualified
This CL makes IsFeatureEnabled(DocumentPolicyFeature...) on Document &
ExecutionContext const qualified so that the migration of
LayoutAnimations from feature policy to document policy can start.

Initially the method is not marked as const because it called
ExecutionContext::AddConsoleMessage which is not marked as const.

Bug: 993790
Change-Id: I98e1b4d210cf3f898c016e004d8d1856dbe41bac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2101707
Commit-Queue: Charlie Hu <chenleihu@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#753720}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5224ec35082196f89f48ee2d0aed56486271679c