Record the update order of the HighestLayers.
We collect three types of the highest layers from subtree for the paint
layer which has overlay overflow control that needs to be reparented.

Since they are all updated individually, the order in which they appear
in the paint layer tree may not be fixed. The update order should be
recorded to ensure that we access them in the proper z-order.

Bug: 1006173
Change-Id: I11aed015a5b14f8a62938274cf7b7c7b58ddf2ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120841
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88e41ff6ca458b2f02af4ca40ea846db30e6579e