SubtreeVisibility: Handle disconnections from viewport observations.
This patch ensures that we gracefully handle the fact that we may
be disconnected when we receive a viewport intersection notification.

This is a consequence of delaying viewport intersection notifications
to the beginning of the next frame.

Note that this isn't a problem right now, but when selection is preventing
locks, then we run into DCHECKs. The patch doing so is upcoming:

https://chromium-review.googlesource.com/c/chromium/src/+/2122857

R=chrishtr@chromium.org

Change-Id: I8449cc6018dd75f3de5962e122b8f79fb250c602
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124975
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754173}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 293f873c1d900488b285edf25e07745f072d2d28