bindings: Make V8SetReturnValue work for bindings::DictionaryBase
We did not have V8SetReturnValue for new IDL dictionaries that does not
take a creation context as an input parameter.
If we tried to call it, another function was called in its overload
resolution.

This CL fixes it.


Bug: 839389
Change-Id: I75c0c96bb3839c166824f2c490cb916dbd345a99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124133
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2305e0f74fc129d2fb5a65bd93b41f8e267f8f62