[RTCInsertableStreams] Add sources to encoded video/audio frames.
This data was intended to be added in a more structured manner
as part of a proper metadata structure to replace the additionalData
field.
However, the final version of the structure has not been decided,
so this CL exposes these fields so that they are available in M83.
All this is behind an experimental flag/origin trial.

Bug: 1052765

Add ssrc

Change-Id: I114171a221b20f2312f043cee84722783c8dbcd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124448
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754444}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b82b6089c38b289c086c6268d1c0d2f76e55f65