[Autofill Auth] Replacing Mojo interface for InternalAuthenticator
Autofill currently uses a Mojo interface to facilitate WebAuthn (and
because it has different security privileges, it requires a slightly
modified implementation.) However, since both Autofill and WebAuthn
exist in the browser process, it is not necessary to have a Mojo pipe.

This CL replaces the Mojo interface with a simple C++ interface, along
with a method "SetEffectiveOrigin".

There is only a Desktop implementation of this interface -- a follow up
CL will create an implementation for Android.


Bug: 949269
Change-Id: I3c209be3132e7ff8482c6783fc4784a38a438a9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2117469
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Commit-Queue: Manas Verma <manasverma@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#754521}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 694e12fb292a297dc01758a1fbce27f90fea1fa0