[CompositeAfterPaint] Composited scrolling of view based on html/body's hint
CSS scrolling properties on html/body apply to the view. For example,
overflow:hidden on html or body actually means the view is not
scrollable. Based on this, using compositing hint on html/body for the
view composited scrolling is reasonable. This seems a better way to hint
composited scrolling of an iframe's content than hinting on the iframe
because this can be controlled by the content document.

Change-Id: I78c3683d1c6f3b610161f5ebfb5cd7d2f6129b11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2125907
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754565}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97a7bfac708557a99ef9f8257bbc719c11e48b36