SubtreeVisibility: Fix-up test assert / comments.
This is a small patch that fixes up the wording of the assert and
some of the comments in the test. Also adds a spacer to ensure that
the test is not flaky under different viewport intersection algorithms.

R=aleventhal@chromium.org

Change-Id: Id56f91cf286aa46dc6a927b465a6dd7e62431380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127508
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754589}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a391155cfe515600b0cbc548d730bae09c0de7a4