Remove DisableCompositingQueryAsserts from ComputeLayerSelection for CAP
On Aura and Android, there is a feature that describes the document/frame
selection to the compositor for touch selection. These coordinates are
layer-relative so that they can later be used to determine whether or not
the start and end of the selection is visible and whether the handles
should be rendered.

This currently relies on Blink being able to query compositing state,
which is not available during PrePaint with CompositeAfterPaint enabled.

This change has us early-out for CAP and leaves a TODO with bug to go
implement this for CompositeAfterPaint

Bug: 1007989

Change-Id: I2876a5bdc847b4e9e8afbbf7b75f4701b387fa23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2120091
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754622}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0a61ecf6bd693aa097811c521b34f1d8c6a2a00