Change virtual test suites to use --enable-gpu-rasterization
The --force-gpu-rasterization flag just ignores the gpu process
capabilities, whereas the --enable-gpu-rasterization flag changes
the capabilities to always allow gpu raster.

We'd like to delete the --force-gpu-rasterization flag as it seems
to provide the same functionality as --enable-gpu-rasterization but in
a much more round-a-bout way.

This has the effect of actually turning on gpu rasterization in the
gpu process in web tests, since --force-gpu-rasterizatoin would not be
passed to the gpu process and web tests would (in the gpu process) then
add the --disable-gpu-rasterization flag in that process. Rebased
web tests reflect that change.

R=kylechar@chromium.org, vmpstr@chromium.org

NOTRY=true

Bug: 866140, 1065893
Change-Id: I049b8909e4a38a7dcb136b2a5e33829e3fe177c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2124890
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754686}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 215a97f8a6b570a0c750340304b09d7bb58a5f2b