Remove DeprecatedEqualIgnoringCase in renderer/core/html
This CL migrates the remaining comparisons involving an ASCII literal
from DeprecatedEqualIgnoringCase to EqualIgnoringASCIICase.

In case of SanitizeValue() in "color_input_type.cc",
Only ASCII (#0-9) and (A-F) characters are involved (not S or K)
in the right hand side, this is effectively not a behavior change.

Also this changes are potentially author-facing, but We’ve optimistically
made them without any additional tests based on Rick Byers’ reasoning
on blink-dev [1].

[1]https://groups.google.com/a/chromium.org/d/msg/blink-dev/sFOpNuQ91UU/3u1HxbnQCQAJ

Bug: 627682

Signed-off-by: DongJun Kim <djmix.kim@samsung.com>
Change-Id: I55afbc42c98ce105a7f94aa572ebb389115f619b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2108240
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Frédéric Wang <fwang@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#754766}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e50427f38b4fa863a10f8dea4edcfab702d57485