Revert find-in-page back to idle task for M83
We changed find-in-page to use normal tasks instead of idle tasks in
crrev.com/c/2041711, but it's currently causing crashes. To ensure
stability in M83, we're reverting the change (but not all of it, we're
keeping the scheduler-side changes to use later - it won't affect the
revert). We'll investigate this and reland after the M83 branch point.

Bug: 1065371
Change-Id: Ib7b0db989910a5c1c0da82eea65af1ab6f617441
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2127998
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#754798}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f43e7c4d79adb0369f11031d7d85733c5f22806a